move: rename RoomTestUtils to TestSampleData
